for labs

+++ src/lab01/ex3.py
@@ -1,6 +1,6 @@
-price = int(input())
-discount = int(input())
-vat = int(input())
+price = float(input())
+discount = float(input())
+vat = float(input())
 base = price * (1 - discount/100)
 vat_amount = base * (vat/100)
 total = base + vat_amount
